resolve 7th sprint
