+++ Sprint_7/helpers.py
@@ -1,5 +1,3 @@
-import datetime
-
 import requests
 from faker import Faker
 
@@ -52,23 +50,6 @@ class DataCreateCourier:
         return data
 
 
-class DataCourier:
-    # валидные данные для регистрации
-    valid_data_login = DataCreateCourier.generating_fake_valid_data_to_create_courier()
-
-    # невалидные данные для регистрации без поля "Login"
-    invalid_data_login_without_login = DataCreateCourier.generating_fake_invalid_data_to_create_courier_without_login_field()
-
-    # невалидные данные для регистрации без поля "Password"
-    invalid_data_login_without_password = DataCreateCourier.generating_fake_invalid_data_to_create_courier_without_password_field()
-
-    # данные несуществующего курьера
-    null_data_login = {
-        "login": "testLogin",
-        "password": "testPassword"
-    }
-
-
 class Courier:
 
     # регистрация в системе с возвратом кода ответа и данных курьера
@@ -91,15 +72,3 @@ class Courier:
         return {"response_text": response.text, "status_code": response.status_code}
 
 
-class DataOrder:
-    # данные для заказа самоката без цвета
-    data = {
-        "firstName": "Тестер",
-        "lastName": "Автотестер",
-        "address": "г.Москва",
-        "metroStation": 2,
-        "phone": "+7 800 555 3535",
-        "rentTime": 4,
-        "deliveryDate": f"{datetime.date.today()}",
-        "comment": "Дай самокат",
-    }
