+++ Sprint_7/tests/test_login_courier.py
@@ -3,7 +3,8 @@ import pytest
 import requests
 
 from Sprint_7.decorators import Step, Check
-from Sprint_7.helpers import DataCourier, Courier
+from Sprint_7.helpers import Courier
+from Sprint_7.data import DataCourier
 from Sprint_7.urls import Urls, Endpoints
